feature: Added test model to the scene
Added "bicycle" test model to the scene and did some basic geometric transformations on it to make it stand upright.
Commented the vertical flip on the models axis to fix broken textures. This model only has a basic diffuse map and exists only to test model placement functionalities, thus it will be removed in a future commit.