feature: Improved lighting shaders
Current lighting shaders have been renamed to object.vs and object.fs as they are intended to calculate light for objects on the scene.
The shaders have been updated to now account for Directional light and Spotlights, as opposed to previously only accounting for Pointlights.
Currently the spotlight and directional light have been disabled, because they shouldn't even exist in the current iteration of the scene, but they will be added once more objects are added.