feature: Added advanced lighting to blending shader and added a cornfield to the scene
Added advanced lighting to the blending shader to make it more realistic. Most of the shader code is similar to the object shader code with a change in texture (with the nonexistence of materials) and the addition of discarding transparent fragments.
Also added a huge cornfield that uses these newly-implemented shaders to the scene.